tee: reject open session and invoke operations with same cancel_id
Prevent the duplicate operation cancel_id from being used concurrently
within one instance of the tee context. Unique cancel_id must be used in
sessions opened and commands invoked from the same tee context fd.

This change rejects duplicate cancel_id when allocating an idr id and is
required for idr_for_each() to make a correct and unique match between a
cancel_id and an idr id.  Otherwise the first idr allocation is always
matched to the detriment of later idr allocations and could result in
sending a cancel request for the wrong operation.